+++ client/package.json
@@ -1,5 +1,6 @@
 {
-  "name": "client",
+  "homepage": "https://JakaPresecnik.github.io/paragoria/",
+  "name": "paragoria",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -8,6 +9,11 @@
     "@testing-library/user-event": "^7.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-icons": "^3.10.0",
+    "react-images": "^1.1.7",
+    "react-photo-gallery": "^8.0.0",
+    "react-player": "^2.3.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
   "scripts": {
