+++ client/public/index.html
@@ -5,11 +5,12 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
+    <meta name="keywords" content="paragoria, paragoria band, paragoria slovenia, paragoria.com, death metal, paragoria metal, decomposition of mind">
+    <meta name="author" content="Jaka Presecnik">
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Official Paragoria site"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
@@ -24,7 +25,13 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>PARAGORIA</title>
+    <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/emailjs-com@2.3.2/dist/email.min.js"></script>
+    <script type="text/javascript">
+       (function(){
+          emailjs.init("user_xQKepu3cq08ACkJ2A99sb"); // Obtain your user ID at the dashboard https://dashboard.emailjs.com/integration
+       })();
+    </script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
