+++ client/src/components/Biography.js
