+++ client/src/img/decomposition_of_mind.jpg
